requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
itypes==1.2.0
packaging==25.0
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
Django>=5.2
gunicorn
psycopg2-binary
dj-database-url
djangorestframework
django-cors-headers
whitenoise
python-dotenv
